Restructure user into component-view and the folder structure
