+++ client/src/components/UserPage/User.js
@@ -1,11 +1,8 @@
 import React, { useState } from 'react'
-import './User.css'
+import UserView from './Views/UserView'
+import './Views/User.css'
 
 import {Redirect} from 'react-router-dom'
-
-import LeftBar from './LeftBar'
-import SubmittedActivity from './Views/SubmittedActivity'
-import ExpandableBukketlist from './ExpandableBukketlist'
 
 const initUser = {
     username: null,
@@ -148,26 +145,3 @@ class UserPage extends React.Component {
 }
 
 export default UserPage
-
-const UserView = ({username, profilePicture, id, description, regDate, loggedInUser, submitted, bukketlists, ...rest}) => {
-    console.log(bukketlists)
-    return (
-        <div>
-            <div className="header">
-                <LeftBar history={rest.history} username={username} profilePicture={profilePicture} id={id} description={description === null ? "" : description} regDate={regDate} loggedInUser={loggedInUser} />
-                <div className="content">
-                        <div><h3>Bukketlists</h3></div>
-                    <div id="bukketlists">
-                        {bukketlists.length === 0 ? `${username} has no bukketlists` : bukketlists.map(x => <ExpandableBukketlist key={x.bid} bukketlist={x} />)}
-                        {/* {bukketlists} */}
-                    </div>
-                    <div><h3>Submitted activities</h3></div>
-                    <div id="submittedactivities">
-                        {submitted.length === 0 ? `${username} has not submitted any activities` : submitted.map(x => <SubmittedActivity key={x.ucaid} text={x.activity} creationDate={x.creationDate}/>)}
-                    </div>
-                </div>
-            </div>
-            
-        </div>
-    )
-}
