+++ client/src/components/UserPage/Views/User.css
@@ -78,8 +78,6 @@
 .content {
     padding: 0 0 0 10px;
     margin: 15px 0 0 15px;
-    /* border-left: solid 1px rgb(210, 210, 210); */
-    /* border: solid; */
     width: 100%;
 }
